Create useTransactionUI hook
